requirements.txt
langchain_ollama
gradio==5.22.0
langchain
langchain_core.prompts
dotenv